# backend/docs/api.yaml
openapi: 3.0.0
info:
  title: GTM E-commerce Backend API
  version: 1.0.0
  description: API documentation for the GTM E-commerce backend.
servers:
  - url: http://localhost:3000

paths:
  /auth/register:
    post:
      summary: Register a new user
      description: Creates a new user account with a unique email and hashed password.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  example: user@example.com
                  description: The user's email address.
                password:
                  type: string
                  minLength: 6
                  example: password123
                  description: The user's password (minimum 6 characters).
      responses:
        '201':
          description: User successfully registered.
          content:
            application/json:
              schema:
                type: object
                properties:
                  user_id:
                    type: integer
                    example: 1
                  email:
                    type: string
                    example: user@example.com
        '400':
          description: Invalid input.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Email and password are required.
        '409':
          description: Email already registered.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Email already registered.
        '500':
          description: Internal server error.

  /auth/login:
    post:
      summary: Authenticate a user
      description: Logs in a user with valid credentials and returns a JWT token.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  example: user@example.com
                  description: The user's email address.
                password:
                  type: string
                  example: password123
                  description: The user's password.
      responses:
        '200':
          description: User successfully authenticated.
          content:
            application/json:
              schema:
                type: object
                properties:
                  user_id:
                    type: integer
                    example: 1
                  email:
                    type: string
                    example: user@example.com
        '400':
          description: Invalid input.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Email and password are required.
        '401':
          description: Invalid email or password.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Invalid email or password.
        '500':
          description: Internal server error.

  /api/settings:
    get:
      summary: List user settings
      description: Retrieves all GTM/GA4 settings for the authenticated user.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of user settings.
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                      example: 1
                    gtm_container_id:
                      type: string
                      example: GTM-XXXXXX
                    ga_measurement_id:
                      type: string
                      example: G-XXXXXX
                    updated_at:
                      type: string
                      format: date-time
                      example: 2025-12-20T12:34:56Z
        '401':
          description: Unauthorized.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Unauthorized.

    post:
      summary: Create a new user setting
      description: Adds a new GTM/GA4 setting for the authenticated user.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - gtm_container_id
                - ga_measurement_id
              properties:
                gtm_container_id:
                  type: string
                  example: GTM-XXXXXX
                  description: The GTM container ID (must start with GTM-).
                ga_measurement_id:
                  type: string
                  example: G-XXXXXX
                  description: The GA4 measurement ID (must start with G-).
      responses:
        '201':
          description: Setting successfully created.
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    example: 1
                  gtm_container_id:
                    type: string
                    example: GTM-XXXXXX
                  ga_measurement_id:
                    type: string
                    example: G-XXXXXX
                  updated_at:
                    type: string
                    format: date-time
                    example: 2025-12-20T12:34:56Z
        '400':
          description: Invalid input.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Invalid GTM Container ID.
        '401':
          description: Unauthorized.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Unauthorized.

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT